src/api_gateway/requirements.txt
flask
requests
redis
retry
psycopg2-binary
